requirements.txt
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
transformers>=4.45.0
torch>=2.2.0
torchaudio>=2.2.0
sounddevice>=0.4.6
numpy>=1.26.0
huggingface_hub>=0.24.0
google-generativeai>=0.8.0
soundfile>=0.13.1
markdown>=3.5.0
sentence-transformers>=2.0.0
